課題11

+++ src/app/admin/posts/[id]/page.tsx
@@ -3,15 +3,19 @@
 
 import { useParams } from "next/navigation";
 import { useRouter } from "next/navigation";
-import { useState, useEffect } from "react";
+import { useState, useEffect, ChangeEvent } from "react";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { Category } from "@/app/_types/Post";
 import { Post, PostCategory } from "@prisma/client";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
+import Image from "next/image";
+import { v4 as uuidv4 } from "uuid";
+import { supabase } from "@/_utils/supabase";
 
 interface Inputs {
   title: string;
   content: string;
-  thumbnailUrl: string;
+  thumbnailImageKey: string;
   categories: Category[];
 }
 
@@ -22,6 +26,14 @@ const PutPost: React.FC = () => {
   const [isLoading, setLoading] = useState(true);
   const [isDeleting, setIsDeleting] = useState(false);
   const [categories, setCategories] = useState<Category[]>([]);
+  const { token } = useSupabaseSession();
+  const [thumbnailImageKey, setThumbnailImageKey] = useState<null | string>(
+    null
+  );
+  const [thumbnailImageUrl, setThumbnailImageUrl] = useState<null | string>(
+    null
+  );
+
   const {
     register,
     handleSubmit,
@@ -33,7 +45,7 @@ const PutPost: React.FC = () => {
     defaultValues: {
       title: "",
       content: "",
-      thumbnailUrl: "",
+      thumbnailImageKey: "",
       categories: [],
     },
   });
@@ -45,15 +57,38 @@ const PutPost: React.FC = () => {
     };
   }
 
+  const getUrl = async (thumbnailImageKey: string) => {
+    if (!thumbnailImageKey) return;
+    setThumbnailImageKey(thumbnailImageKey);
+    // アップロード時に取得した、thumbnailImageKeyを用いて画像のURLを取得
+    const {
+      data: { publicUrl },
+    } = await supabase.storage
+      .from("post_thumbnail")
+      .getPublicUrl(thumbnailImageKey);
+    setThumbnailImageUrl(publicUrl);
+  };
+
   useEffect(() => {
+    if (!token) return;
+
     const fetcher = async () => {
       setLoading(true);
-      const resp = await fetch(`/api/admin/posts/${id}`, { method: "GET" });
+      const resp = await fetch(`/api/admin/posts/${id}`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
+      });
       const { post }: PostResponse = await resp.json();
+
+      getUrl(post.thumbnailImageKey);
+
       reset({
         title: post.title,
         content: post.content,
-        thumbnailUrl: post.thumbnailUrl,
+        thumbnailImageKey: post.thumbnailImageKey,
         categories: post.postCategories.map(postCategory => {
           return {
             id: postCategory.category.id,
@@ -64,7 +99,7 @@ const PutPost: React.FC = () => {
       setLoading(false);
     };
     fetcher();
-  }, [reset]);
+  }, [reset, token]);
 
   interface CategoryResponse {
     status: number;
@@ -72,24 +107,35 @@ const PutPost: React.FC = () => {
   }
 
   useEffect(() => {
+    if (!token) return;
     const fetcher = async () => {
       const categoryResp = await fetch("/api/admin/categories", {
         method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token, // 👈 Header に token を付与
+        },
       });
       const { data }: CategoryResponse = await categoryResp.json();
       setCategories(data);
     };
     fetcher();
-  }, []);
+  }, [token]);
 
   if (isLoading) return <div>読み込み中...</div>;
 
   const handleDeletePost = async () => {
+    if (!token) return;
+
     const confirmDelete = window.confirm("削除してもよろしいですか？");
     if (!confirmDelete) return;
     setIsDeleting(true);
     const prams = {
       method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: token,
+      },
       body: JSON.stringify({ id: parseInt(id) }),
     };
     const data = await fetch(endPoint, prams);
@@ -100,26 +146,28 @@ const PutPost: React.FC = () => {
       router.push("/admin/posts");
     } else {
       window.alert("削除に失敗しました");
-      console.log(status);
     }
   };
 
   const onSubmit: SubmitHandler<Inputs> = async data => {
-    const categoryIds = data.categories.map(item => item.id);
-    const prams = {
-      method: "PUT",
-      body: JSON.stringify({
-        id: parseInt(id),
-        title: data.title,
-        content: data.content,
-        thumbnailUrl: data.thumbnailUrl,
-        categoryIds: categoryIds,
-      }),
-    };
+    if (!token) return;
     try {
-      const resp = await fetch("/api/admin/posts/[id]", prams);
+      const prams = {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token, // 👈 Header に token を付与
+        },
+        body: JSON.stringify({
+          id: parseInt(id),
+          title: data.title,
+          content: data.content,
+          thumbnailImageKey: thumbnailImageKey,
+          categoryIds: data.categories.map(item => item.id),
+        }),
+      };
+      const resp = await fetch(`/api/admin/posts/${id}`, prams);
       const contents = await resp.json();
-      console.log(contents);
       if (contents.status === 200) {
         window.alert("登録に成功しました");
         router.push("/admin/posts");
@@ -128,7 +176,6 @@ const PutPost: React.FC = () => {
       }
     } catch (e) {
       if (e instanceof Error) {
-        console.log(e.message);
         window.alert("登録に失敗しました");
       }
     }
@@ -147,6 +194,31 @@ const PutPost: React.FC = () => {
       const category = categories.find(c => c.id === id);
       setValue("categories", [...watch("categories"), category as Category]);
     }
+  };
+  const handleImageChange = async (
+    event: ChangeEvent<HTMLInputElement>
+  ): Promise<void> => {
+    if (!event.target.files || event.target.files.length == 0) {
+      // 画像が選択されていないのでreturn
+      return;
+    }
+    const file = event.target.files[0]; // 選択された画像を取得
+    const filePath = `private/${uuidv4()}`; // ファイルパスを指定
+    // Supabaseに画像をアップロード
+    const { data, error } = await supabase.storage
+      .from("post_thumbnail") // ここでバケット名を指定
+      .upload(filePath, file, {
+        cacheControl: "3600",
+        upsert: false,
+      });
+    // アップロードに失敗したらエラーを表示して終了
+    if (error) {
+      alert(error.message);
+      return;
+    }
+    // data.pathに、画像固有のkeyが入っているので、thumbnailImageKeyに格納する
+    setThumbnailImageKey(data.path);
+    getUrl(data.path);
   };
 
   return (
@@ -181,13 +253,25 @@ const PutPost: React.FC = () => {
         </label>
         <div className="w-full">
           <input
-            type="text"
-            id="thumbnailUrl"
-            className="border border-gray-300 rounded-lg p-4 w-full mb-4"
+            type="file"
+            id="thumbnailImageKey"
+            className="pt-4 w-full mb-4"
             disabled={isSubmitting || isDeleting}
-            {...register("thumbnailUrl")}
+            accept="image/*"
+            onChange={handleImageChange}
           ></input>
         </div>
+        {thumbnailImageUrl && (
+          <div className="mt-2">
+            <Image
+              src={thumbnailImageUrl}
+              alt="thumbnail"
+              width={400}
+              height={400}
+            />
+          </div>
+        )}
+
         <label htmlFor="title" className="w-[240px]">
           カテゴリー
         </label>
